Added feature to init the database on the first run.
When the user runs pyTalkManager for the first time, a dialogbox asks the user for a location to save the file. It then takes the location and saves it to the config.ini file along with switching the firsttimerunning key to False. A few other addtions were made along the way as well, for example, adding the configSet function for setting values in config.ini.