Changed the names of the widgets.
Some of the widgets had names that didn't follow the convention I had
for the lineEdit widgets. All the widgets now have a standard convention
of first half of widget name and what the widget records separated by a
underscore. For example, the LineEdit for first name becomes line_f_name.