Edited the Talk Manager window began working on issue #102.
Added an Import button to the dialog to allow the user to import a list
of talks if no talk outlines are in the database.

Problem: For some strange reason, the QT4 Designer installed on my
laptop is not working properly. I can not see or edit the widget's
properties from the Property Editor. I will have to fix that problem
before continuing with this branch.

See issue #102 for more information about this branch.