Began to implement the Outline feature in pyTalkManager.
Started to implement the outline manager and features to pyTalkManager
but it's not finished.

Added tool tips to Outline-Manager and added function for the Import button to auto disable when database isn't empty.

Changed the columns from title & subject to number & title.

Implemented importing of outlines from a file to the database.

Added populate_list function to populate outlines in database.

Added visibility column to Talk table.

Stated implementing the Delete feature in the Outline window.

Also other minor edits such as expending the docstrings for DB.py.

Added ability to delete. Done.